Remove promise returning from nick/join/part
This feature was implemented naively. It works fine when the nick/join/part command is successful, but there are many errors that aren't handled. If an error happens, the promise never rejects and the listener for the promise remains, causing a memory leak.

It's not really possible to 100% be sure when a command is rejected, as some error replies are ambiguous as to which command it is a reply to. Alas, this is the nature of IRC.